# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Aiexec
  version: 1.6.4
paths:
...
  /api/v1/build/{flow_id}/flow:
    post:
      tags:
      - Chat
      summary: Build Flow
      description: "Build and process a flow, returning a job ID for event polling.\n\
        \nThis endpoint requires authentication through the CurrentActiveUser dependency.\n\
        For public flows that don't require authentication, use the /build_public_tmp/flow_id/flow\
        \ endpoint.\n\nArgs:\n    flow_id: UUID of the flow to build\n    background_tasks:\
        \ Background tasks manager\n    inputs: Optional input values for the flow\n\
        \    data: Optional flow data\n    files: Optional files to include\n    stop_component_id:\
        \ Optional ID of component to stop at\n    start_component_id: Optional ID\
        \ of component to start from\n    log_builds: Whether to log the build process\n\
        \    current_user: The authenticated user\n    queue_service: Queue service\
        \ for job management\n    flow_name: Optional name for the flow\n    event_delivery:\
        \ Optional event delivery type - default is streaming\n\nReturns:\n    Dict\
        \ with job_id that can be used to poll for build status"
      operationId: build_flow_api_v1_build__flow_id__flow_post
      security:
      - OAuth2PasswordBearer: []
      - API key query: []
      - API key header: []
      parameters:
      - name: flow_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Flow Id
      - name: stop_component_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Stop Component Id
      - name: start_component_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Start Component Id
      - name: log_builds
        in: query
        required: false
        schema:
          type: boolean
          default: true
          title: Log Builds
      - name: flow_name
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Flow Name
      - name: event_delivery
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/EventDeliveryType'
          default: polling
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Body_build_flow_api_v1_build__flow_id__flow_post'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/build_public_tmp/{flow_id}/flow:
    post:
      tags:
      - Chat
      summary: Build Public Tmp
      description: "Build a public flow without requiring authentication.\n\nThis\
        \ endpoint is specifically for public flows that don't require authentication.\n\
        It uses a client_id cookie to create a deterministic flow ID for tracking\
        \ purposes.\n\nThe endpoint:\n1. Verifies the requested flow is marked as\
        \ public in the database\n2. Creates a deterministic UUID based on client_id\
        \ and flow_id\n3. Uses the flow owner's permissions to build the flow\n\n\
        Requirements:\n- The flow must be marked as PUBLIC in the database\n- The\
        \ request must include a client_id cookie\n\nArgs:\n    flow_id: UUID of the\
        \ public flow to build\n    background_tasks: Background tasks manager\n \
        \   inputs: Optional input values for the flow\n    data: Optional flow data\n\
        \    files: Optional files to include\n    stop_component_id: Optional ID\
        \ of component to stop at\n    start_component_id: Optional ID of component\
        \ to start from\n    log_builds: Whether to log the build process\n    flow_name:\
        \ Optional name for the flow\n    request: FastAPI request object (needed\
        \ for cookie access)\n    queue_service: Queue service for job management\n\
        \    event_delivery: Optional event delivery type - default is streaming\n\
        \nReturns:\n    Dict with job_id that can be used to poll for build status"
      operationId: build_public_tmp_api_v1_build_public_tmp__flow_id__flow_post
      parameters:
      - name: flow_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Flow Id
      - name: stop_component_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Stop Component Id
      - name: start_component_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Start Component Id
      - name: log_builds
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          default: true
          title: Log Builds
      - name: flow_name
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Flow Name
      - name: event_delivery
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/EventDeliveryType'
          default: polling
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Body_build_public_tmp_api_v1_build_public_tmp__flow_id__flow_post'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/upload/{flow_id}:
    post:
      tags:
      - Base
      summary: Create Upload File
      description: 'Upload a file for a specific flow (Deprecated).


        This endpoint is deprecated and will be removed in a future version.'
      operationId: create_upload_file_api_v1_upload__flow_id__post
      deprecated: true
      parameters:
      - name: flow_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Flow Id
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_create_upload_file_api_v1_upload__flow_id__post'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/aiexec__api__v1__schemas__UploadFileResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/flows/{flow_id}:
    get:
      tags:
      - Flows
      summary: Read Flow
      description: Read a flow.
      operationId: read_flow_api_v1_flows__flow_id__get
      security:
      - OAuth2PasswordBearer: []
      - API key query: []
      - API key header: []
      parameters:
      - name: flow_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Flow Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlowRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - Flows
      summary: Update Flow
      description: Update a flow.
      operationId: update_flow_api_v1_flows__flow_id__patch
      security:
      - OAuth2PasswordBearer: []
      - API key query: []
      - API key header: []
      parameters:
      - name: flow_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Flow Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FlowUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlowRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Flows
      summary: Delete Flow
      description: Delete a flow.
      operationId: delete_flow_api_v1_flows__flow_id__delete
      security:
      - OAuth2PasswordBearer: []
      - API key query: []
      - API key header: []
      parameters:
      - name: flow_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Flow Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/flows/public_flow/{flow_id}:
    get:
      tags:
      - Flows
      summary: Read Public Flow
      description: Read a public flow.
      operationId: read_public_flow_api_v1_flows_public_flow__flow_id__get
      parameters:
      - name: flow_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Flow Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlowRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/flows/batch/:
    post:
      tags:
      - Flows
      summary: Create Flows
      description: Create multiple new flows.
      operationId: create_flows_api_v1_flows_batch__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FlowListCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/FlowRead'
                type: array
                title: Response Create Flows Api V1 Flows Batch  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
      - API key query: []
      - API key header: []
...
  /api/v1/flows/download/:
    post:
      tags:
      - Flows
      summary: Download Multiple File
      description: Download all flows as a zip file.
      operationId: download_multiple_file_api_v1_flows_download__post
      requestBody:
        content:
          application/json:
            schema:
              items:
                type: string
                format: uuid
              type: array
              title: Flow Ids
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
      - API key query: []
      - API key header: []
...
components:
  schemas:
...
    Body_build_flow_api_v1_build__flow_id__flow_post:
      properties:
        inputs:
          anyOf:
          - $ref: '#/components/schemas/InputValueRequest'
          - type: 'null'
        data:
          anyOf:
          - $ref: '#/components/schemas/FlowDataRequest'
          - type: 'null'
        files:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Files
      type: object
      title: Body_build_flow_api_v1_build__flow_id__flow_post
    Body_build_public_tmp_api_v1_build_public_tmp__flow_id__flow_post:
      properties:
        inputs:
          anyOf:
          - $ref: '#/components/schemas/InputValueRequest'
          - type: 'null'
        data:
          anyOf:
          - $ref: '#/components/schemas/FlowDataRequest'
          - type: 'null'
        files:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Files
      type: object
      title: Body_build_public_tmp_api_v1_build_public_tmp__flow_id__flow_post
...
    EventDeliveryType:
      type: string
      enum:
      - streaming
      - direct
      - polling
      title: EventDeliveryType
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: api/v1/login
    API key query:
      type: apiKey
      in: query
      name: x-api-key
    API key header:
      type: apiKey
      in: header
      name: x-api-key